Support braceless namespaces
Added several tricks to make braceless namespaces transparantly work
just like regular namespaces. Not ideal but guaranteed to work. It's
also a lot beter than resolving the FQNs directly in semantic-fetch-tags.

The semanticdb/typecache integration with phptags has been removed as
well in this commit, just temporarily so we can add it back
later. The problem there was that it didn't allow semantic to search in
the project database, and used the fallback to phptags very early.

In future commits I plan to:

 * leverage the project databases (looking up tags in current and open
   buffers, for which this commit is a precursor - this almost works)
 * leverage the ede-composer component (like includes in semantic but
   not quite like that, unfortunately)
 * fall back to the omniscience database (this happens by default, we
   just need to restore the phptags backend for semanticdb)
 * make use-statements work for superclasses (that concept doesn't exist
   in c++ so semantic doesn't understand it)
 * separate phptags as omniscience database (like gnu global) from
   phptags as system symbol database (like javap). They work a bit
   differently in semantic and in phptags as well, so we better threat
   them differently as well.