Work-in-progress restoration of phptags backend
It's not yet optional, but technically it's an augmentation of the
default behaviour so it's easy to make it optional and still have the
same logic. For example: if you just open the buffers manually, you get
the same functionality as without phptags. That's what I'm aiming for
here.

So this branch is on it's way to do the same as current master, but
without a hacky/overridden semantic-analyze-current-context. Don't
worry, enough hacks and todo's in there.

In this branch I'm also trying to figure out where and how the EDE
component will fit in.